get latest version of FarColorer from colorer repo

+++ misc/nightly/colorer.sh
@@ -3,12 +3,18 @@
 function bcolorer {
   BIT=$1
   PLUGIN=FarColorer
-  COLORER_VERSION=1.4.4
+  COLORER_VERSION=$(curl -fsLJ 'https://raw.githubusercontent.com/colorer/FarColorer/master/version4far.txt')
+  if [ -z "$COLORER_VERSION" ]; then
+    echo "Failed to get Colorer version"
+    return 1
+  fi
+
+  echo "Download FarColorer ${COLORER_VERSION}"
   COLORER_PLATFORM=$2
   COLORER_FILE_NAME=FarColorer.${COLORER_PLATFORM}.v${COLORER_VERSION}.7z
   
   rm -f ${COLORER_FILE_NAME}
-  curl -fsLJO -o ${COLORER_FILE_NAME} https://github.com/colorer/FarColorer/releases/download/v${COLORER_VERSION}/${COLORER_FILE_NAME}
+  curl -fsLJO -o ${COLORER_FILE_NAME} 'https://github.com/colorer/FarColorer/releases/download/v${COLORER_VERSION}/${COLORER_FILE_NAME}'
   if [ ! -e ${COLORER_FILE_NAME} ]; then
     echo "Can't find ${COLORER_FILE_NAME}"
     return 1
